% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macamts-package.R
\docType{package}
\name{macamts-package}
\alias{macamts}
\alias{macamts-package}
\title{macamts: A subset of the R package macam for time series analysis}
\description{
macamts contains a subset of the R functions implemented in macam. The functions are mainly for time series analysis.
}
\keyword{internal}
